requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf==3.20.5
langchain_experimental
mysql-connector-python
pymysql
torch
sentence-transformers==2.2.2
chromadb